✨ Add table groups concept to dbStructure
