+++ frontend/packages/db-structure/src/schema/dbStructure.ts
@@ -1,6 +1,8 @@
 import * as v from 'valibot'
 
 // Export these schema definitions
+export const tableGroupNameSchema = v.string()
+
 export const columnNameSchema = v.string()
 
 export const tableNameSchema = v.string()
@@ -77,9 +79,22 @@ export type Tables = v.InferOutput<typeof tablesSchema>
 const relationshipsSchema = v.record(relationshipNameSchema, relationshipSchema)
 export type Relationships = v.InferOutput<typeof relationshipsSchema>
 
+// Schema for table group
+export const tableGroupSchema = v.object({
+  name: v.string(),
+  tables: v.array(tableNameSchema),
+  comment: v.nullable(v.string()),
+})
+
+export type TableGroup = v.InferOutput<typeof tableGroupSchema>
+
+const tableGroupsSchema = v.record(tableGroupNameSchema, tableGroupSchema)
+export type TableGroups = v.InferOutput<typeof tableGroupsSchema>
+
 export const dbStructureSchema = v.object({
   tables: tablesSchema,
   relationships: relationshipsSchema,
+  tableGroups: v.optional(tableGroupsSchema),
 })
 
 export type DBStructure = v.InferOutput<typeof dbStructureSchema>
