+++ frontend/packages/db-structure/src/schema/index.ts
@@ -1,4 +1,4 @@
-export { dbStructureSchema } from './dbStructure.js'
+export { dbStructureSchema, tableGroupSchema } from './dbStructure.js'
 export type {
   Column,
   Columns,
@@ -11,6 +11,7 @@ export type {
   Indexes,
   ForeignKeyConstraint,
   Cardinality,
+  TableGroup,
 } from './dbStructure.js'
 export {
   aColumn,
@@ -22,7 +23,5 @@ export {
 export {
   applyOverrides,
   dbOverrideSchema,
-  tableGroupSchema,
-  tableGroupNameSchema,
 } from './overrideDbStructure.js'
-export type { TableGroup, DBOverride } from './overrideDbStructure.js'
+export type { DBOverride } from './overrideDbStructure.js'
